requirements.txt
numpy
matplotlib
pandas
nltk==3.5
sklearn
keras==2.4.3
tensorflow==2.3.0
tensorflow-hub==0.9.0
bert-for-tf2==0.14.5
slackclient==2.8.0
slackeventsapi==2.2.1
Flask==1.1.2
certifi==2020.6.20
xlrd
jupyter